[AC-561] Move delete-organization-dialog into its own folder
