+++ apps/web/src/app/organizations/settings/index.ts
@@ -1,2 +1,2 @@
 export * from "./organization-settings.module";
-export { DeleteOrganizationDialogComponent } from "./delete-organization-dialog.component";
+export { DeleteOrganizationDialogComponent } from "./components/delete-organization-dialog/delete-organization-dialog.component";
