+++ apps/web/src/app/organizations/settings/organization-settings.module.ts
@@ -4,7 +4,7 @@ import { LooseComponentsModule, SharedModule } from "../../shared";
 import { PoliciesModule } from "../policies";
 
 import { AccountComponent } from "./account.component";
-import { DeleteOrganizationDialogComponent } from "./delete-organization-dialog.component";
+import { DeleteOrganizationDialogComponent } from "./components/delete-organization-dialog/delete-organization-dialog.component";
 import { OrganizationSettingsRoutingModule } from "./organization-settings-routing.module";
 import { SettingsComponent } from "./settings.component";
 import { TwoFactorSetupComponent } from "./two-factor-setup.component";
